+++ src/remotes/zsimulator/zsimremote.ts
@@ -322,7 +322,7 @@ export class ZSimRemote extends DzrpRemote {
 		}
 
 		// Check for keyboard
-		const zxKeyboard = zsim.zxKeyboard || zsim.zxInterface2Joy;
+		const zxKeyboard = (zsim.zxKeyboard !== 'none') || zsim.zxInterface2Joy;
 		if (zxKeyboard) {
 			this.zxKeyboard = new ZxKeyboard(this.ports);
 		}